randomGenerator: Localise globalScalar01 to injection models
Use of this function is discouraged. It is expensive and introduces
non-randomness in all cores other then the master. It is better practice
to keep a generator synchronised by calling it the same number of times
on the different processes. It has been removed from the interface of
the random generator and localised to the lagrangian injection models,
which are the only place left in the code that use it.