polyMesh: Subdirectory for mesh objects
Poly-mesh mesh objects are now in a subdirectory of the polyMesh
directory. The cpuLoad object has been made a polyMesh object, rather
than an fvMesh object, in order to remove the dependence on the finite
volume library. This has, in turn, removed the base lagrangian library's
dependence on finite volume.