Time: Changed FatalError in the time precision update to Warning
Due to rounding to the nearest value it is possible for the updated time name to
predate the previous time name after a precision change, for example as a
consequence of a sudden reduction in time-step.  This is not an error and should
not cause the run to stop but the user is warned about the temporary anomaly in
the printing of time.