Function1s::Table: Column control in foam format files
Foam-format table files can now have an arbitrary number of columns. For
example, a file containing the following can be read:

    (
        (0 1 (1 2 3) 4 (5 6 7 8 9 10))
        (1 2 (2 3 4) 5 (6 7 8 9 10 11))
        (2 3 (3 4 5) 6 (7 8 9 10 11 12))
    )

The columns used in the table are then specified using a 'columns'
entry. For example:

    UMean
    {
        type    table;
        file    "UMean.foam";
        columns (0 2);
    }

This says to use column #0 as the x-axis and column #2 as the UMean
values. Note that column #2 contains vectors. Vectors and tensors and
similar complex types are considered to be a single column and are
identified as such by their parentheses. This is in contrast to the CSV
reader in which the components of these types are read from different
columns.

By default, columns #0 and #1 are assumed for the x-axis and the
values, respectively. This means for simple two-column tables, the
behaviour remains the same.