sixDoFMotion, sixDoFAcceleration: Motion components in separate functions
These functions now specify motion components (i.e., translation,
rotation, acceleration, angular rotation and acceleration) in separate
functions, rather than as a single combined Tuple2 or Vector of vectors
type. This means that the different components can have different
functional forms, and that unit conversion of the components is now
possible. It also makes it much easier to deactivate a component by
setting it to a constant value of zero or similar.

If the motion is specified as tabulated data then this can in general
still be read from a single tabulated data file by specifying different
columns in the different motion component functions. See
tutorials/incompressibleVoF/sloshingTank3D6DoF for an example of this.