XiFluid: Separated the Xi model and associated sub-models into run-time selectable classes
The Xi model choices of fixed, algebraic and transport can now be selected via
the model selection of the flameWrinkling sub-dictionary of
combustionProperties, e.g.:

flameWrinkling
{
    model           transport;

    equilibrium
    {
        model           Gulder;
    }

    generationRate
    {
        model           KTS;
    }
}

in which the equilibrium and generation-rate sub-models are also specified.  All
models now have defaults provided for all coefficients for user convenience.

This system is more convenient and flexible to use and allows for additional
model choices to be added by the user to the selection tables.