topoSet: Added setAndPointToFaceZone
This adds another way of defining the orientation of a face zone
constructed from a face set. In this method, a point is provided that is
considered to be behind (or inside) the face zone. The point must be
within the mesh, and the provided face set must unambiguously separate
cells that are connected to the point from those that are not. This
means that the set must be closed or must span the entire width of the
domain, from boundary to boundary.

Whilst more restrictive than setToAndNormalToFaceZone, this method
allows sets to be reliably converted into zones even when it is not
possible to specify a normal direction that applies to all the faces;
for example a set representing a closed surface, or just one with
substantial variation in its surface normals.