codedBase: Expand the code strings so the context dictionary entries are not needed on restart
With this changes any substitutions in the code from the outer context are
expanded so

    inlet
    {
        type            codedFixedValue;
        name            swirl;
        Ux              1;
        code
        #{
            const vector axis(1, 0, 0);
            vectorField v(2*this->patch().Cf() ^ axis);
            v.replace(vector::X, $<scalar>Ux);
            operator==(v);
        #};
        value           $internalField;
    }

is now written as

    inlet
    {
        type            codedFixedValue;
        value           nonuniform List<vector>
        .
        .
        .
        name            swirl;
        code            #{
            const vector axis(1, 0, 0);
            vectorField v(2*this->patch().Cf() ^ axis);
            v.replace(vector::X, Foam::read<scalar>("1"));
            operator==(v);
        #};
    }

where the Ux entry has been expanded.  This avoids the need to cache the context
dictionary just in case the code uses entries from it.  Before this change the
above codedFixedValue would run from the start but would not restart due to the
missing Ux entry.