fvConstraints: Removed the cached coefficients sub-dictionary
The coefficients sub-dictionary is now looked up from the parent IOdictionary as
required in the constructor and read member function to avoid the need to cache
a copy of the dictionary.  This changes ensures the context of any coefficient
lookup is correct which provides the error messaging system with sufficient
information about the global nature of the fvConstraints dictionary to issue
messages from the master processor only when running in parallel, significantly
improving the readability of syntax error messages.