surfaceFieldValue: Added the ability to specify multiple patches
The 'select' control can now take the value 'patches' in which case a
'patches' control will be used to specify a list of patches, rather than
just a single patch.

The 'patches' and 'patch' controls now also both support wildcards.

For example to recover the flow rate through multiple outlets:

    outletFlowRate
    {
        type            surfaceFieldValue;
        libs            ("libfieldFunctionObjects.so");

        writeControl    timeStep;
        writeInterval   1;

        writeFields     false;

        select          patch;
        patch           ".*Outlet";

        // or ...
        //select          patches;
        //patches         (outlet1 outlet2 ".*Outlet");

        fields          (phi);

        operation       sum;
    }

If 'writeFields' is set to true, this function now also writes a single
surface containing all the fields. Previously it wrote a separate
surface for each field, as well as an additional surface containing no
field values. This was absurdly wasteful.