createNonConformalCouples: Pass additional entries to created patches
Any additional entries in the settings for a couple are now forwarded to
the non-conformal patches being constructed. At present, the only
additional optional entry that is available is the 'reMapAfterMove'
control which prevents re-computing the intersection when it is known
that the patches do not move. But, if additional controls are added in
future then this forwarding will also support them.

For example, to set 'reMapAfterMove' for a non-conformal mapped wall
patch, in the system/createNonConformalCouplesDict:

    fields          no;

    nonConformalCouples
    {
        fluidSolidStationary
        {
            regions         (fluid solid);
            patches         (nonCoupleStationaryFluid nonCoupleStationarySolid);
            reMapAfterMove  no; // <-- !!!
        }
    }

Or, equivalently:

    fields          no;

    nonConformalCouples
    {
        fluidSolidStationary
        {
            owner
            {
                region          fluid;
                patch           nonCoupleStationaryFluid;
                reMapAfterMove  no; // <-- !!!
            }
            neighbour
            {
                region          solid;
                patch           nonCoupleStationarySolid;
                reMapAfterMove  no; // <-- !!!
            }
        }
    }