nonConformalCyclic, mappedPatchBase: moveUpdate control
This entry chooses the condition that triggers re-calculation of the
intersection/mapping following mesh motion. It can take three values;
'always', 'detect', or 'never'.

A setting of 'always' will re-calculate the intersection/mapping,
'never' will not, and 'detect' will compare the new and old points and
re-calculate if there is any difference

The comparison that 'detect' does has an expense associated with it, so
it should only really be used if the patch is both moving and static at
different times. If the patch is always in motion or is fully
stationary, then it is more efficient to specify 'always' or 'never'.

'always' is the default, as it is safest to assume that motion
always affects the intersection/mapping.

This control replaces the 'reMapAfterMove', but 'reMapAfterMove' entries
are still read for backwards compatibility, and will select 'always' if
true, and 'never' if false.