MomentumTransportModels: Removed the printCoeffs clutter
The printCoeffs functionality only existed in the MomentumTransportModels, it
was never propagated to any other physical sub-models in OpenFOAM, rather
limiting its usefulness.  Also the implementation was rather complex requiring
care in coding when applying it to derived classes, adding class type name
checking to avoid duplicate calls and required a copy of the coefficients
dictionary which meant that any errors in the momentumTransport file would be
printed from all processors for parallel runs.

All this clutter has been removed in preference for the completely general
writeOptionalEntries switch which enables printing off all defaulting settings,
switches, coefficents etc. for ALL sub-models:

    InfoSwitches
    {
        writeOptionalEntries 1;
    }

While the output of this may not be as convenient as that from printCoeffs it is
very simple to use, completely general and does not require any specific coding
making it much easier to maintain which is a critical requirement for all
open-source code development given the very limited number of companies prepared
to contribute to maintenance funding