sampledSet: Updated handing of mesh changes
Rather than requiring the calling code to delete and reconstruct sampledSets if
the mesh changes the standard mesh change call-backs are now provided

    // Mesh changes

        //- Update for mesh motion
        virtual bool movePoints();

        //- Update topology using the given map
        virtual void topoChange(const polyTopoChangeMap&);

        //- Update from another mesh using the given map
        virtual void mapMesh(const polyMeshMap&);

        //- Redistribute or update using the given distribution map
        virtual void distribute(const polyDistributionMap&);

to handle the changes necessary for each type of mesh change.  This also avoids
the need to cache the configuration dictionary in the sampledSets
functionObject, significantly simplifying the code and maintenance.