mappedPatchBaseBase: Fix transform specification options
The 'transformIsNone/defaultTransformIsNone' boolean has been replaced
by a ternary enumeration which denotes whether the transform is none,
whether it is none by default but can be specified as something else, or
whether it must be specified.

Broadly, transforms must be specified for non-wall mapped patches, they
may be specified for mapped walls but will otherwise default to none,
and they cannot be specified for extruded walls and film patches.

An error is now also generated when using a patchToPatch method if one
or more faces are not coupled.