populationBalance: Added size group boundary and source conditions
Two boundary and two source conditions have been added for size-group
fractions. These are called singleSizeGroup and distributionSizeGroup.
singleSizeGroup sets a value of one in a single group, identified by its
index, and a value of zero in all other groups. distributionSizeGroup
sets values corresponding to a given distribution.

Example specifications:

    <patchOrSourceName>
    {
        type        singleSizeGroup;
        index       0;
        value       $internalField;
    }

    <patchOrSourceName>
    {
        type        distributionSizeGroup;
        distribution
        {
            type        RosinRammler;
            min         1 [mm];
            max         15 [mm];
            d           14 [mm];
            n           2 [];
        }
        value       $internalField;
    }

(note that the syntax is identical for boundary and source conditions)

These conditions reduce the need for the user to explicitly create
multiple size-group fraction field files. They can be applied to the
0/fDefault.<phase> file and will then apply consistently to multiple
size groups.