createNonConformalCouples: Do not stitch meshes in constant/meshes
When createNonConformalCouples is used on a mesh within constant/meshes,
non-coformal patches are still created, but the intersection is no
longer computed and these patches are therefore not populated with
faces.

This prevents unnecessary recalculation of the intersection, and it
means that region meshes within constant/meshes can now have couples
created and be decomposed and reconstructed independently of the regions
that it is connected to.

The added region independence means that constructing a case with
mesh-to-mesh mapping and non-conformal couples no longer involves moving
meshes to and from the primary location in order to construct couplings
and decompose.