solvers::solid: Removed the dubious diffusion number control
A diffusion number based purely on the diffusivity and cell geometry is not a
reliable way of controlling the time-step for solid thermal diffusion due to the
sensitivity to cell shape and independence of the thermal gradients.

It would be possible to create a time-step control based on the actual flux of
the field, more similar to the way in which the Courant number is evaluated, but
the estimated rate of change of the field due to diffusion would need to be
normalised, i.e. the user would need to supply a normalisation level, the
maximum amount the field would be allowed to change within one
time-step. However, given that the diffusion is solved implicitly and there are
no stability issues relating to the solution, this control only relates to
temporal accuracy and it is not clear what kind of control is really needed for
this; more research is required. In the meantime the dubious diffusion number
control from the solid solver has been removed.

Resolved bug-report https://bugs.openfoam.org/view.php?id=4154