functionObjects::timeControl: Moved startTime, endTime into Foam::timeControl
Now functionObjects execute at the optionally specified startTime (which
defaults to the case beginTime) and at intervals starting at that time rather
than the case beginTime and at intervals starting then.  This new behaviour is
more obvious with respect to the controls provided but it does mean that care
needs to be exercised if the user wishes the dumps from the functionObject to
coincide with field write times if startTime is specified, i.e. in this case
startTime should be specified as one of the write times.