unitConversion: Fixed format of non-named units
Non-named unit conversions are now written in a single set of square
brackets, with colons used to separate the sets of exponents and
multipliers. So, for example, a named unit conversion of [rpm] gets
converted to a non-named [0 0 -1 0 0 0 0 : 0 1 : 0.10472] (i.e.,
[time^-1 : angle : 0.10472]).

Previously, multiple sets of square brackets were used to separate the
sections, so the above example would have been written as
[0 0 -1 0 0 0 0][0 1][0.10472]. The problem with this is that '][' is
not distinguished in the token parser from '] ['. Combined with the fact
that the later sections are optional, this makes reading pairs or lists
of these objects impossible. The new format is, by comparison,
unambiguously delimited in all contexts.

This change may break the restart of a case in which unit conversion
entries have been written out. It is unlikely to require modification to
an initial configuration as named unit conversion entries (such as [rpm]
and [MPa] and [l/min]) are unaffected.