momentumTransportModel::devTau,devSigma: now return the surface stress as a surfaceVectorField
This change ensures that fvc::div(devTau()) and fvc::div(devSigma()) are
consistent with the divDevTau(U) and divDevSigma(U) matrix evaluation and the
wallShearStress and forces functionObjects also provide stresses and forces
consistent with the momentum solution.  This consistency is particularly
important when evaluating momentum, force and energy balances for the flow
solution.