test/boundaryLayer/boundary.*/fvSchemes: Updated laplacian(nuEff,U) -> laplacian(interpolate(nuEff),U)
In order to ensure the implicit and explicit terms in the stress are consistent
and consistent between devTau and divDevTau nuEff is now interpolated separately
so that all the terms use the same interpolation.  To specify a different
interpolation practice than linear for nuEff the entry should now be put in
interpolationSchemes rather than part of the laplacian specification.