nonConformalMappedWall: User-controllable owner and neighbour
The "owner" side of a non-conformal mapped wall interface is now defined
by the order in which the patches are specified within the
createNonConformalCouplesDict. This is consistent with non-conformal
cyclics and means that the owner and neighbour can be indirectly
specified by the user.

The owner patch is projected onto the neighbour, so its geometry
changes, whether as the neighbour patch's geometry remains unaffected.
This difference can be important if one side has a much finer or more
layered mesh than the other, in which case, it may be desirable not to
project and alter this side. This can now be achieved by making this
patch the neighbour by putting it second in the pair of patches in the
createNonConformalCouplesDict.