fvMeshStitcher: Error/warn about quality measures
An error is now generated if the critical quality measures (openness and
volume conservation error) exceed a value that might be expected as a
result of round-off error. A substantial value of one of these measures
would indicate a bug in the stitcher.

A warning is also now generated if the (non-critical) projected volume
fraction exceeds a value; currently 30%. This indicates a substantial
change in the face areas of NCC-adjacent cells, which might then affect
stability, under-relaxation, etc..., settings. A high value of this
measure might indicate that NCC might be better configured with the
owner and neighbour reversed. Projected volume fraction is reduced if
the side of the interface with smaller or more finely layered cells is
the neighbouring patch (i.e., is specified second on the command line or
in the patches/regions entries for createNonConformalCouples).