functionObjects::removeObjects: Rename and packaging
This function is a rename of removeRegisteredObject. This makes it clear
that it can operate on multiple objects, it removes the unnecessary
clarification that the object is registered (all objects modified by
function objects have to be registered), and it makes the function
correspond more consistently with the writeObjects function.

A packaged function has also been added. Its usage is trivial. At its
simplest it takes only the objects that need removing as arguments.

For example, to interpolate the turbulent kinetic energy field to the
faces, average it on a face zone, and then remove the interpolated field
again:

    #includeFunc surfaceInterpolate(executeControl=timeStep, executeInterval=1, k)
    #includeFunc faceZoneAverage(faceZone=fz1, surfaceInterpolate(k))
    #includeFunc removeObjects(executeControl=timeStep, executeInterval=1, surfaceInterpolate(k))