simpleRushtonMRF/NCC tutorials: cylindrical tank with a Rushton turbine.
The tutorials generate the mesh using blockMesh, mirrorMesh, topoSet and
createBaffles.  The Allmesh script to generate the mesh includes the
following options:

-b | -bafflesOff   do not include baffles
-g | -graded       apply grading to resolve boundary layers on tank wall
-i | -impeller <n> number of impeller blades 4 (default) or 6
-r | -refineRotating  apply refinement around the rotatingZone
-s | -size <size>  mesh size, 1-4, 1=~68k cells (default), 4=~4M cells

There are two versions of the tutorial - simpleRushtonMRF and
simpleRushtonNCC - which simulate the rotating stirrer using multiple
rotating frames (MRF) and non-conformal coupling (NCC), respectively.

simpleRushtonMRF is a steady-state simulation, while simpleRushtonNCC is
transient with mesh motion.  The Allrun script for simpleRushtonNCC
includes an option to start the simulation from a existing solution,
typically the steady one from the simpleRushtonMRF case.