Matrix: Added std::initializer_list constructors
To simplify the initialised construction of SquareMatrix, SymmetricSquareMatrix,
RectangularMatrix and DiagonalMatrix constructors from std::initializer_list and
std::initializer_list<std::initializer_list> are now provided.  For example a
square matrix can either be constructed from single list of all the element
values in row order:

    SquareMatrix<scalar> M
    {
        4, 1, 1, 12, 37, 2, -16, -43, 98
    };

or from separate rows of element values:

    SquareMatrix<scalar> M
    {
        {-3.0, 10.0, -4.0},
        {2.0, 3.0, 10.0},
        {2.0, 6.0, 1.0}
    };

If a RectangularMatrix is constructed from a single list of all the element
values the number of elements in the rows and columns must be provided, e.g.

    RectangularMatrix<scalar> M
    (
        3, 2,
        {4, 1, 1, 12, 37, 2}
    );

but if separate rows of element values are provided the number of elements in
each row is available from those lists and need not be provided as separate
arguments to the constructor, i.e.

    RectangularMatrix<scalar> M
    {
        {4, 1},
        {1, 12},
        {37, 2}
    };

The Test-Matrix.C test application has been updates to demonstrate this
development.