fvModels: phaseChange: Support both single- and multi-specie changes
This is a change from the previous structure in which single- and
multi-specie phase changes were separate derivations of the phase change
base class. The derivation-based distinction is not flexible enough to
allow for models which switch between single- and multi-specie phase
changes based on user input.

An additional benefit of this change is that the coefficientPhaseChange
model now supports multiple species.