nonConformalCHT: new template case for CHT using NCC interface
The nonConformalCHT template case is designed to help create a conjugate
heat transfer (CHT) simulation in which fluid and solid regions are
coupled with non-conformal coupled (NCC) interfaces.  It includes an
Allmesh script which provides example commands to generate the meshes
for the fluid and solid regions using the 'snappyHexMeshConfig' utility,
and to generate NCC interfaces using 'createNonConformalCouples'.  The
remaining workflow to configure and run the CHT simulation is similar to
the 'singleFluidCHT' template case.