Foam::solver: Added support for PIMPLE flow, thermophysics and models controls in ALL solver modules
This required separating momentum and thermophysical transport prediction and
correction in all solver modules by

Adding

    //- Predict the momentum transport
    virtual void momentumTransportPredictor() = 0;

    //- Predict thermophysical transport
    virtual void thermophysicalTransportPredictor() = 0;

and replacing

    //- Correct the momentum and thermophysical transport modelling
    virtual void postCorrector() = 0;

with

    //- Correct the momentum transport
    virtual void momentumTransportCorrector() = 0;

    //- Correct the thermophysical transport
    virtual void thermophysicalTransportCorrector() = 0;

This allowed the controls for transport prediction and correction to be moved
from the solver modules into foamRun and foamMultiRun along with the PIMPLE
flow, thermophysics and models controls from multiphaseEuler.

Now in the PIMPLE controls in fvSolution the optional controls

    flow            yes;
    thermophysics   yes;
    models          yes;

are available to switch on/off these parts of the solution, in particular for
complex CHT simulations in which the thermal time-scales are much longer than
the fluid time-scales and it is advantageous to run the thermal system on a
frozen flow field after it has evolved and then coupling the solution again
after the thermal system has evolved.  These switches can be changed at run-time
allowing the flow and thermal systems to be switched on/off during the run.