Semi-implicit MULES: Added optional limiter convergence test and approximate non-iterative form
The CMULES (correction MULES) limiter is used to limit the fluxes of the
explicit corrections applied to an implicit prediction for large
time-step/Courant number simulations, particularly for pseudo-steady VoF
simulations.  The number of MULES limiter iterations needed to ensure
boundedness of the solution is dependent on the Courant number, typically ~3 +
Courant number which is inconvenient to specify and adjust.  Now there is an
optional convergence criterion which can be specified so that the number of
limiter iterations adjusts automatically, for example in the
incompressibleVoF/planingHullW3 tutorial case:

        MULESCorr       yes;
        nLimiterIter    100;
        MULEStolerance  2e-4;
        alphaApplyPrevCorr  yes;

here the maximum number of MULES limiter iterations is 100 but the iteration
stops when the changes in normalised limited flux reduce below MULEStolerance
and this value is closely related to the unboundedness of the subsequent
solution, i.e. the value 2e-4 ensures that the unboundness is < 2e-4.  With the
above settings the planingHullW3 runs about 10% faster due to the overall
reduction in the number of limiter iterations required, just enough to ensure
boundedness and of course easier to specify and the settings are independent on
the Courant number of the simulation.  The convergence of the limiter can be
optionally printed by adding

DebugSwitches
{
    MULES           1;
}

to controlDict, see the incompressibleVoF/planingHullW3 case.

Additionally an optional simplified, approximate non-iterative form of the
limiter is provide which is selected by setting nLimiterIter to 1.  In this form
the positive flux limiter is calculated assuming the negative flux limiter is 0
and vice versa, thus removing the interdependence between the positive and
negative flux limiters and the need for iteration.  In this form of
limiter is more aggressive that the standard MULES limiter which introduces the
absolute minimum amount of limiter needed to guarantee boundedness, and hence
produces more smearing and lower accuracy solutions but is substantially cheaper
to compute.  However when applied iteratively to correct an implicit solution
the additional smearing from the non-iterative approximate limiter is
successively corrected out, so it can be applied successfully to large Courant
number simulations where temporal accuracy is compromised to some extent
anyway.  The incompressibleVoF/DTCHull, incompressibleVoF/DTCHullMoving and
tutorials/incompressibleVoF/DTCHullWave cases have been updated to use this new
form of limiter and the solutions are almost identical to those using the
standard limiter but ~10% cheaper to run.

Unfortunately this non-iterative approximate limiter is too diffusive for the
incompressibleVoF/planingHullW3 case, causing slight air-bleed under the hull
from the interface smearing near the bow of the hull, hence the iterative MULES
limiter is used with the convergence tolerance specified.