tmp: Restored check for assignment to null
The check for assignment to null was removed in order to be able to
store tmp objects in a Pair, which as a derivation of FixedList required
tmp objects to be null constructable and assignable. This check can now
be restored as it is compatible with Pair now that Pair no longer
inherits from FixedList and instead stores its member data directly.