dimensionedConstants: Use the unit conversion system to convert to different unit sets
This allows the fundamental constants to be removed from the
etc/controlDict file and hard-coded with a defined unit set. The
standardsisation of the constant set, particularly with respect to moles
and standard pressure and temperature, means that simulations involving
thermodynamics are now possible in the USCS unit set (or others). It
also means that non-SI unit sets are now required to provide a minimal
set of conversions to SI units in order to convert the fundamental
constants.