MULES: Improved iterative algorithm to guarantee boundedness irrespective of limiter convergence
The limiter iterative algorithm in both explicit MULES and semi-implicit CMULES
has been rewritten to start from 0 and increase towards the converged limit
rather than starting from 1 and reducing towards the converged limit.  This
profound change ensures that the limiter will provide a bounded solution of the
transported scalar irrespective of number of iterations or convergence of the
limiter, even if only one iteration is performed, but of course the accuracy of
the solution is improved by better convergence of the limiter.

At low Courant numbers 2 or 3 iterations of the limiter are usually sufficient
for convergence and accuracy but with the semi-implicit formulation at large
Courant numbers the number of iterations required scales with the Courant
number, ~3 + Courant number is usually sufficient but no longer strictly
necessary for boundedness if an approximate solution is all that is required,
for example to evolve the system as fast as possible before a more accurate
lower time-step simulation is performed.

To simplify the specification of the limiter convergence there is an optional
convergence criterion which can be specified for either explicit and
semi-implicit MULES so that the number of limiter iterations adjusts
automatically, for example in the incompressibleVoF/planingHullW3 tutorial case:

        MULESCorr       yes;
        alphaApplyPrevCorr  yes;
        nLimiterIter    20;
        MULEStolerance  1e-2;

here the maximum number of MULES limiter iterations is 20 but the iteration
stops when the changes in normalised limited flux reduce below MULEStolerance
1e-2 which is sufficient for accuracy of the interface capturing and boundedness
is guaranteed irrespective of the value of MULEStolerance.

The convergence of the limiter can be optionally printed by adding

DebugSwitches
{
    MULES           1;
}

to controlDict, see the incompressibleVoF/planingHullW3 case.