VoF solvers: nAlphaSubCycles is now a Function1 of the alpha Courant number
The explicit MULES sub-cycling control nAlphaSubCycles is now a Function1 of the
alpha Courant number which allows a very convenient way to specify the number of
sub-cycles independent of the time-step the case is run with while maintaining
stability of the explicit solution of the phase-fraction.  For example with the
following entry in fvSolution:

solvers
{
    alpha.water
    {
        nAlphaSubCycles polynomial (0 2);
    }
.
.
.

nAlphaSubCycles is evaluated as twice the current maximum alpha Courant number
rounded upwards to an integer.  For better temporal accuracy of the
phase-fraction transport the scaling factor can be increased for example in some
tutorial cases

    nAlphaSubCycles polynomial (0 4);

is used.

This useful development applies to all four VoF solvers: incompressibleVoF,
compressibleVoF, incompressibleMultiphaseVoF and compressibleMultiphaseVoF.