OldTimeField: Use calculated for old-time patch types
Previously full copies of the boundary conditions were made, but only
the values were then overwritten when the time was incremented. This
left old data in place, potentially resulting in incorrect results of
operations performed on old-time fields.

Either a full copy of all data should be done using the reset method or
only the values should be retained (in calculated patch fields). The
latter option has been chosen in this instance as this should be the
more efficient choice as it copies and stores less data overall.