dynamicMesh: Interoperability improvements
A number of fixes have been made to ensure interoperability between
various combinations of motion, refinement, distribution and
second-order time-schemes. A test script has been placed in the
incompressibleVoF/rotatingCube tutorial to provide a convenient way to
verify that all supported combinations continue to function.

The combination that remains unsupported is motion, refinement and a
second-order time scheme. This is because second-order time schemes
require information regarding the old-time mesh motion (i.e., mesh flux
and old-time cell volumes), but it is not clear how this information can
be mapped as a result of topology change in such a way as to preserve
conservation of the mesh volume. If this combination is selected then an
error will be generated.