gradient/mixed energy/unburntEnthalpy boundary conditions: Specify values on unmapped faces
These conditions are mapped by making the approximation that the
unmapped faces are zero-gradient/adiabatic. This is appropriate for
these cases as unmapped faces typically repesent a region of the
boundary that was previously absent, and is therefore sensibly
represented by zero heat flux. This choice maintains this state until
the boundary condition is properly re-evaluated, at which point its
parameters are inferred from the corresponding temperature condition.